Trivial modification(debug)

+++ deploy/backup.sh
@@ -2,7 +2,7 @@
 
 cd /srv/kumanodocs/
 SQL_BACKUP_PATH=deploy/backup
-if [ ! -e ${SQL_BACKUP_PATH}] ; then
+if [ ! -e ${SQL_BACKUP_PATH} ] ; then
     mkdir ${SQL_BACKUP_PATH}
 fi
 if test "07" = $(date +"%d") -o "22" = $(date +"%d"); then
@@ -13,10 +13,12 @@ if test "07" = $(date +"%d") -o "22" = $(date +"%d"); then
     rm ${SQL_SEED}.gz
 else
     SOURCE_BACKUP=${SQL_BACKUP_PATH}/kumanodocs-database-backup-seed-$(date +"%Y-%m").sqlite3
-    if [ ! -e ${SOURCE_BACKUP}]; then
-        cp db.sqlite3 ${SOURCE_BACKUP}
-        rm ${SQL_BACKUP_PATH}/kumanodocs-database-backup-diff*
-    fi
     DIFF_BACKUP=${SQL_BACKUP_PATH}/kumanodocs-database-backup-diff-$(date +"%Y-%m-%d").sqlite3
+    if [ ! -e ${SOURCE_BACKUP} ]; then
+        cp db.sqlite3 ${SOURCE_BACKUP}
+        if [ -e ${SQL_BACKUP_PATH}/kumanodocs-database-backup-diff* ]; then
+            rm ${SQL_BACKUP_PATH}/kumanodocs-database-backup-diff*
+        fi
+    fi
     sqldiff ${SOURCE_BACKUP} db.sqlite3 > ${DIFF_BACKUP}
 fi
